Refactor nozomi and first tests
:KyaruThumbsUp:
